Separate out upload utilities, adjust tests
